+++ build/kernel6.6-build-on-ubuntu22.04.sh
@@ -56,7 +56,9 @@ if ls "${BUILDER_DIR}/kernel-6.6/"*.patch >/dev/null 2>&1; then
   git am ${BUILDER_DIR}/kernel-6.6/*.patch
 fi
 # config kernel
-cp -a ${BUILDER_DIR}/kernel-6.6/config-6.6 .config
+if [ -f ${BUILDER_DIR}/kernel-6.6/config-6.6 ];then
+  cp -a ${BUILDER_DIR}/kernel-6.6/config-6.6 .config
+fi
 make ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- olddefconfig
 make ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- -j$(nproc) Image
 make ARCH=arm64 CROSS_COMPILE=aarch64-linux-gnu- -j$(nproc) modules
